+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 * Core: add `requestVoid` method to network client.
 * Core: default error parser now ignores `.explicitlyCancelled` (for cancelled requests).
 * Core: result "cancellation" now uses Swift's built-in `CancellationError` type if available.
+* Core: `TaggedIdentifiable` protocol now conforms to Swift's own `Identifiable` protocol.
 * XcodeGen: added a template for shared frameworks.
 
 ### Bug Fixes